Backport from trunk:
unicodeobject.c 2.169
stringobject.c 2.189

Fix warnings on 64-bit platforms about casts from pointers to ints.
Two of these were real bugs.

+++ Objects/unicodeobject.c
@@ -5642,7 +5642,8 @@ PyObject *PyUnicode_Format(PyObject *format,
                              "unsupported format character '%c' (0x%x) "
                              "at index %i",
                              (31<=c && c<=126) ? (int)c : '?', 
-                             (int)c, (fmt -1 - PyUnicode_AS_UNICODE(uformat)));
+                             (int)c,
+                             (int)(fmt -1 - PyUnicode_AS_UNICODE(uformat)));
                 goto onError;
 	    }
 	    if (sign) {
